fixed integer types

+++ dbt_subprojects/dex/models/trades/ronin/platforms/tamadotmeme_v1_ronin_base_trades.sql
@@ -21,11 +21,11 @@ with trades AS (
     CAST(DATE_TRUNC('day', evt_block_time) as date) AS block_date, -- Extract day from block time
     evt_block_time AS block_time, -- Keep original block time
     evt_block_number AS block_number, -- Keep block number
-    case when isBuy=True then cast(amountOut as double)  -- If it's a buy, amountOut is the bought amount
-         else cast(amountIn as double)                 -- Otherwise, amountIn is the bought amount
+    case when isBuy=True then cast(amountOut as uint256)  -- If it's a buy, amountOut is the bought amount
+         else cast(amountIn as uint256)                 -- Otherwise, amountIn is the bought amount
     end AS token_bought_amount_raw,
-    case when isBuy=True then cast(amountIn as double)  -- If it's a buy, amountIn is the sold amount
-         else cast(amountOut as double)                 -- Otherwise, amountOut is the sold amount
+    case when isBuy=True then cast(amountIn as uint256)  -- If it's a buy, amountIn is the sold amount
+         else cast(amountOut as uint256)                 -- Otherwise, amountOut is the sold amount
     end AS token_sold_amount_raw,
     case when isBuy=True then token                    -- If it's a buy, the 'token' column holds bought token
          else {{wron_token_address}}                  -- Otherwise, it's wRON
